perf(db-postgres): skip pagination overhead if `limit: 0` is passed (#12261)
This improves performance when querying data in Postgers / SQLite with
`limit: 0`. Before, unless you additionally passed `pagination: false`
we executed additional count query to calculate the pagination. Now we
skip this as this is unnecessary since we can retrieve the count just
from `rows.length`.

This logic already existed in `db-mongodb` -
https://github.com/payloadcms/payload/blob/1238f87b0e0b0f3bd3200c811be35aa22d0d5301/packages/db-mongodb/src/find.ts#L114-L124