fix(db-mongodb): documents not showing in folders with `useJoinAggregations: false` (#14155)
### What?

Documents were not showing in folders in admin views. This has actually
not been working since v3.48.0 (#12763), just haven't had the time to
fix 😅

### Why?

The admin folder view was generating a join query like the following:

```ts
joins: {
  documentsAndFolders: {
    limit: 100000000,
    sort: 'name',
    where: {
      or: [
        {
          and: [
            { relationTo: { equals: 'payload-folders' } },  // <-- breaking here
            {
              or: [{ folderType: { in: ['posts'] } }, { folderType: { exists: false } }],
            },
          ],
        },
        {
          and: [{ relationTo: { equals: 'posts' } }],
        },
      ],
    },
  },
},
```

The `extractRelationToFilter()` function in `resolveJoins.ts` was not
correctly processing `and` and `or` queries. It was breaking on the
first `relationTo` found in the array, thus only returning folders and
not other documents.

### Fixes

`extractRelationToFilter()` now handles `and`/`or` where queries in
joins correctly