fix: allow slugField to accept localized argument and fixed slug generation with custom field names (#14234)
Previously you had to use the overrides to enable localization for the
`slugField` however it's a very common request that slugs are localized
so we've added it to the args.

Now you can do:
```ts
slugField({ localized: true })
```

There was also a bug relating to using custom field names, even though
we accept it the hook would still look for `slug`, this has been fixed.

Todo:
- [x] Add e2e tests for general field and localized field