fix(richtext-lexical): bulk upload drawer rendered at the wrong depth if already in a drawer (#14472)
This wraps the Lexical editor in the BulkUploadProvider which matches
the Upload Field. Now, when the lexical upload feature calls the
useBulkUpload hook it has access to the correct edit depth.

Alternative approaches might be to modify the BulkUploadProvider to pass
in the depth directly but this seemed simplest and matches the Upload
field.

Fixes #14426

---------

Co-authored-by: Alessio Gravili <alessio@gravili.de>