fix: localized block fields (#14587)
Fixes https://github.com/payloadcms/payload/issues/13663

Previously https://github.com/payloadcms/payload/pull/13974 was merged
but it didn't actually fix the issue. I updated the tests to make them
fail before making the fix in this PR.

The visual issue stemmed from to autosave not passing the
`fallback-locale=null` param. So the update function would pass the
default fallback locale into afterRead and data for the nullified field
would come back.

The data part of that issue was previously solved by adding an array to
other locales if not defined. However that clutters the db with unused
data. This can be fixed by including the `formatBaseSchema` in the block
schema which (for localized fields) sets `schema.sparse = true`, so the
data is only added when it is saved.