fix(plugin-multi-tenant): fix infinite sync tenant network calls (#14604)
Fixes an issue where selecting a tenant, logging out and back in would
cause the network to be cluttered with sync tenant calls. This happened
because the initialValue (initialCookie) for the tenant does not clear
when the user logs out. This PR calls router.refresh on log out so the
initialValue is reset after the cookie is deleted from logging out.