fix: remove init of transaction for global doc access (#14693)
When getting the globals doc permissions we were creating a transaction.
This was causing transaction errors to occur where the session was
closed already due to a promise.All() during
[renderDocument](https://github.com/payloadcms/payload/blob/7b74c33d13fa9983156000dbc6594e571a8080a5/packages/next/src/views/Document/index.tsx#L170)

This change makes it so we no longer create the transaction as is normal
for all other db read operations.