fix: add beforeDocumentControls to globals generate importmap (#15036)
### What?

When adding `beforeDocumentControls` as a custom component to a global,
it does not get picked up by the import map, even when running
`generate:importmap`.

### Why?

`beforeDocumentControls` is missing from `addToImportMap` in
`iterateGlobals`.

### How?

Adds `beforeDocumentControls` to `addToImportMap` in `iterateGlobals`.

While doing this I've used the opportunity to add `Description` to the
globals custom component documentation as it was missing.

Fixes #15035

Co-authored-by: Riley Langbein <rilrom2@hotmail.com>